refact!: Merge TOTP dialog controllers
