+++ config/areas/users/dialogs.php
@@ -13,7 +13,7 @@ use Kirby\Panel\Controller\Dialog\UserChangePasswordDialogController;
 use Kirby\Panel\Controller\Dialog\UserChangeRoleDialogController;
 use Kirby\Panel\Controller\Dialog\UserCreateDialogController;
 use Kirby\Panel\Controller\Dialog\UserDeleteDialogController;
-use Kirby\Panel\Controller\Dialog\UserTotpDisableDialogController;
+use Kirby\Panel\Controller\Dialog\UserTotpDialogController;
 use Kirby\Panel\Controller\Dialog\UserWebauthnDialogController;
 
 return [
@@ -45,9 +45,9 @@ return [
 		'pattern' => 'users/(:any)/delete',
 		'action'  => UserDeleteDialogController::class
 	],
-	'user.totp.disable' => [
-		'pattern' => 'users/(:any)/totp/disable',
-		'action'  => UserTotpDisableDialogController::class
+	'user.totp' => [
+		'pattern' => 'users/(:any)/totp',
+		'action'  => UserTotpDialogController::class
 	],
 	'user.webauthn' => [
 		'pattern' => 'users/(:any)/webauthn',
