--- src/Panel/Controller/Dialog/UserTotpDisableDialogController.php
@@ -1,105 +1,0 @@
-<?php
-
-namespace Kirby\Panel\Controller\Dialog;
-
-use Kirby\Cms\App;
-use Kirby\Cms\Find;
-use Kirby\Cms\User;
-use Kirby\Exception\InvalidArgumentException;
-use Kirby\Exception\PermissionException;
-use Kirby\Panel\Ui\Dialog;
-use Kirby\Panel\Ui\Dialog\FormDialog;
-use Kirby\Toolkit\Escape;
-
-/**
- * Controls the Panel dialog to disable TOTP auth for a user
- *
- * @package   Kirby Panel
- * @author    Nico Hoffmann <nico@getkirby.com>
- * @link      https://getkirby.com
- * @copyright Bastian Allgeier
- * @license   https://getkirby.com/license
- * @since     6.0.0
- */
-class UserTotpDisableDialogController extends UserDialogController
-{
-	public static function factory(string|null $id = null): static
-	{
-		return new static($id ? Find::user($id) : App::instance()->user());
-	}
-
-	public function load(): Dialog
-	{
-		$currentUser = $this->kirby->user();
-		$submitBtn   = [
-			'text'  => $this->i18n('disable'),
-			'icon'  => 'protected',
-			'theme' => 'negative'
-		];
-
-		// admins can disable TOTP for other users without
-		// entering their password (but not for themselves)
-		if (
-			$currentUser->isAdmin() === true &&
-			$currentUser->is($this->user) === false
-		) {
-			$name = $this->user->name()->or($this->user->email());
-
-			return new FormDialog(
-				text: $this->i18n('login.totp.disable.admin', [
-					'user' => Escape::html($name)
-				]),
-				submitButton: $submitBtn,
-			);
-		}
-
-		// everybody else
-		return new FormDialog(
-			fields: [
-				'password' => [
-					'type'     => 'password',
-					'required' => true,
-					'counter'  => false,
-					'label'    => $this->i18n('login.totp.disable.label'),
-					'help'     => $this->i18n('login.totp.disable.help'),
-				]
-			],
-			submitButton: $submitBtn
-		);
-	}
-
-	/**
-	 * Removes the user's TOTP secret when the dialog is submitted
-	 */
-	public function submit(): array
-	{
-		$password = $this->request->get('password');
-
-		try {
-			if ($this->kirby->user()->is($this->user) === true) {
-				$this->user->validatePassword($password);
-			} elseif ($this->kirby->user()->isAdmin() === false) {
-				throw new PermissionException(
-					message: 'You are not allowed to disable TOTP for other users'
-				);
-			}
-
-			// Remove the TOTP secret from the account
-			$this->user->changeSecret('totp', null);
-
-			return [
-				'message' => $this->i18n('login.totp.disable.success')
-			];
-		} catch (InvalidArgumentException $e) {
-			// Catch and re-throw exception so that any
-			// Unauthenticated exception for incorrect passwords
-			// does not trigger a logout
-			throw new InvalidArgumentException(
-				key: $e->getKey(),
-				data: $e->getData(),
-				fallback: $e->getMessage(),
-				previous: $e
-			);
-		}
-	}
-}
