+++ tests/Panel/Controller/Dropdown/UserSettingsDropdownControllerTest.php
@@ -54,85 +54,4 @@ class UserSettingsDropdownControllerTest extends TestCase
 		$options    = $controller->load();
 		$this->assertCount(9, $options);
 	}
-
-	public function testLoadWithTotpEnable(): void
-	{
-		$this->app = $this->app->clone([
-			'options' => [
-				'auth' => [
-					'methods' => [
-						'password' => ['2fa' => true]
-					]
-				]
-			]
-		]);
-
-		$this->app->impersonate('test@getkirby.com');
-
-		$user       = $this->app->user('test');
-		$controller = new UserSettingsDropdownController($user);
-		$options    = $controller->load();
-		$this->assertCount(10, $options);
-		$this->assertSame('/account/totp/enable', $options[7]['dialog']);
-	}
-
-	public function testLoadWithTotpDisable(): void
-	{
-		$this->app = $this->app->clone([
-			'options' => [
-				'auth' => [
-					'methods' => [
-						'password' => ['2fa' => true]
-					]
-				]
-			]
-		]);
-
-		$this->app->impersonate('test@getkirby.com');
-
-		$user = $this->app->user('test')->changeTotp('ABCDEFGHIJKLMNOPQRSTUVWXYZ234567');
-		$controller = new UserSettingsDropdownController($user);
-		$options    = $controller->load();
-		$this->assertCount(10, $options);
-		$this->assertSame('/account/totp/disable', $options[7]['dialog']);
-	}
-
-	public function testTotpMode(): void
-	{
-		$user       = $this->app->user('test');
-		$controller = new UserSettingsDropdownController($user);
-		$this->assertNull($controller->totpMode());
-
-		$this->app = $this->app->clone([
-			'options' => [
-				'auth' => [
-					'methods' => [
-						'password' => ['2fa' => true]
-					]
-				]
-			]
-		]);
-
-		$user       = $this->app->user('test');
-		$controller = new UserSettingsDropdownController($user);
-		$this->assertNull($controller->totpMode());
-
-		$this->app->impersonate('test@getkirby.com');
-
-		$user       = $this->app->user('test');
-		$controller = new UserSettingsDropdownController($user);
-		$this->assertSame('enable', $controller->totpMode());
-
-		$user = $user->changeTotp('ABCDEFGHIJKLMNOPQRSTUVWXYZ234567');
-		$controller = new UserSettingsDropdownController($user);
-		$this->assertSame('disable', $controller->totpMode());
-
-		$this->app->impersonate('admin@getkirby.com');
-		$controller = new UserSettingsDropdownController($user);
-		$this->assertSame('disable', $controller->totpMode());
-
-		$user = $user->changeTotp(null);
-		$controller = new UserSettingsDropdownController($user);
-		$this->assertNull($controller->totpMode());
-	}
 }
